requirements.txt
cryptography==1.7.1
enum34==1.1.6
idna==2.2
ipaddress==1.0.17
Pillow==5.0.0
pyasn1==0.1.9
pycrypto==2.6.1
pygobject==3.22.0
pynput==1.3.7
python-rtmidi==1.1.0
python-xlib==0.20
pyxdg==0.25
SecretStorage==2.3.1
six==1.11.0
spidev==3.2
